Update initial value of H 17
The value in the source code broke the logic for determining if a
routine had already been loaded into memory. This updated value seems to
avoid that issue on the first load. After the first routine is loaded,
teh values are all defined by JOSS and seem to work correctly.